Change location pipeline autocompletion and use gitlab dummy
