+++ tests/components/test_completion.py
@@ -2,7 +2,10 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from nf_core.components.components_completion import autocomplete_modules, autocomplete_subworkflows, autocomplete_pipelines
+from nf_core.components.components_completion import autocomplete_modules, autocomplete_subworkflows
+
+from ..utils import GITLAB_NFTEST_BRANCH, GITLAB_URL
+
 
 class DummyParam:
     # Minimal mock object for Click parameter (not used in the function)
@@ -14,14 +17,23 @@ class DummyCtx:
         self.obj = obj
         self.params = params if params is not None else {}
 
+
 def test_autocomplete_modules():
-    ctx = DummyCtx()
+    ctx = DummyCtx(
+        obj={
+            "modules_repo_url": GITLAB_URL,
+            "modules_repo_branch": GITLAB_NFTEST_BRANCH,
+            "modules_repo_no_pull": True,
+        }
+    )
     param = DummyParam()
-    completions = autocomplete_modules(ctx, param, "bcf")
+    completions = autocomplete_modules(ctx, param, "samt")
 
     values = [c.value for c in completions]
-    assert "bcftools/call" in values
+    assert "samtools/stats" in values
+    assert "samtools/idxstats" in values
     assert "fastqc" not in values
+
 
 @patch("nf_core.components.components_completion.CompletionItem")
 @patch("nf_core.components.components_completion.ModuleList")
@@ -85,14 +97,20 @@ def test_autocomplete_modules_missing_argument(capfd):
 
 
 def test_autocomplete_subworkflows():
-    ctx = DummyCtx()
+    ctx = DummyCtx(
+        obj={
+            "modules_repo_url": GITLAB_URL,
+            "modules_repo_branch": GITLAB_NFTEST_BRANCH,
+            "modules_repo_no_pull": True,
+        }
+    )
     param = DummyParam()
-    completions = autocomplete_subworkflows(ctx, param, "utils")
-    print(completions)
+    completions = autocomplete_subworkflows(ctx, param, "bam_stats")
 
     values = [c.value for c in completions]
-    assert "utils_nextflow_pipeline" in values
-    assert "vcf_gather_bcftools" not in values
+    assert "bam_stats_samtools" in values
+    assert "bam_sort_stats_samtools" not in values
+
 
 @patch("nf_core.components.components_completion.CompletionItem")
 @patch("nf_core.components.components_completion.SubworkflowList")
@@ -158,46 +176,5 @@ def test_autocomplete_subworkflows_missing_argument():
 
     with pytest.raises(TypeError) as exc_info:
         autocomplete_subworkflows(ctx, param)  # Missing 'incomplete' argument
-
-    assert "missing 1 required positional argument" in str(exc_info.value)
-
-def test_autocomplete_pipelines():
-    ctx = DummyCtx()
-    param = DummyParam()
-    completions = autocomplete_pipelines(ctx, param, "next")
-
-    values = [c.value for c in completions]
-    assert "nextflow-io/hello" in values
-    assert "nf-core/rnasek" not in values
-
-@patch("nf_core.components.components_completion.Workflows")
-def test_autocomplete_pipelines_mocked(mock_workflows_class):
-    # Mock instance
-    mock_instance = mock_workflows_class.return_value
-
-    # Mock local and remote workflows
-    mock_instance.local_workflows = [MagicMock(full_name="awesome/localpipeline")]
-    mock_instance.remote_workflows = [MagicMock(full_name="awesome-remote"), MagicMock(full_name="other-remote")]
-
-    ctx = DummyCtx()
-    param = DummyParam()
-
-    completions = autocomplete_pipelines(ctx, param, "awesome")
-
-    # Extract values from CompletionItem
-    values = [c.value for c in completions]
-
-    # Assertions
-    assert "awesome/localpipeline" in values
-    assert "awesome-remote" in values
-    assert "other-remote" not in values
-
-
-def test_autocomplete_pipelines_missing_argument(capfd):
-    ctx = DummyCtx()
-    param = DummyParam()
-
-    with pytest.raises(TypeError) as exc_info:
-        autocomplete_pipelines(ctx, param)  # Missing 'incomplete' argument
 
     assert "missing 1 required positional argument" in str(exc_info.value)
